Added Python module to work around strings/list issue.
Essentially, it's not possible to call getattr(...) in the Python 3 builtins,
since the strings are interpreted as lists, getattr barfs. This code wraps the
builtin for getattr as well as making calls to a callable attr. It does so by
accepting binaries from Erlang, which get converted to Python bytes and can
easily be converted to strings once in Python.